packages/jawikt/pyproject.toml
[project]
name = "jawikt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "mkpoli", email = "git@mkpo.li" }]
requires-python = ">=3.13"
dependencies = [
    "requests>=2.32.3",
    "wikitext",
    "wiktionary",
    "zstandard>=0.23.0",
]

[project.optional-dependencies]
mecab = ["mecab-python3>=1.0.10"]
janome = ["janome>=0.5.0"]
vibrato = ["vibrato>=0.2.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
wikitext = { workspace = true }
wiktionary = { workspace = true }
